move dag tests_fast into tests, move analysis.utils into analysis.embeddings, remove duplicated metrics in baselines
